requirements.txt
pandas>=1.0
openai>=1.0
python-dotenv>=1.0